Join current query to another omop table

Simple wrapper for join functions to join an existing query to another table in the omop cdm (or any other of the same source). Include the following two lines at the top of your script after setting the schema and connection

options(con.default.value = connection_variable_name)

options(schema.default.value = cdm_schema_variable_name)

omop_join(
  data,
  table,
  type,
  con = getOption("con.default.value"),
  schema = getOption("schema.default.value"),
  ...
)

Arguments

data

sql query from dbplyr/dplyr. this function works in pipes!

table

the omop table (or other table in your schema) you wish to join

type

the type of join. use types available in dplyr: left, right, inner, anti, full etc.

con

defaults to the connection you set with options(). does not need to be specified

schema

defaults to the schema you set with options(). does not need to be specified.

...

arguments passed on to the join function. e.g., by = "person_id"

Value

more sql query info stuff

Examples

options(con.default.value = con)
#> Error in options(con.default.value = con): object 'con' not found
options(schema.default.value = cdm_schema)
#> Error in options(schema.default.value = cdm_schema): object 'cdm_schema' not found
obs_tbl |>
  omop_join("person", type = "left", by = "person_id")
#> Error in omop_join(obs_tbl, "person", type = "left", by = "person_id"): Remember to set the connection and schema defaults. see ?omop_join for details