Get occcurances of of a concept set from ATLAS for a given cohort

Get occcurances of of a concept set from ATLAS for a given cohort

pull_concept_set(
  cohort,
  concept_set_id,
  start_date,
  end_date,
  name = "has_concept",
  min_n = NULL,
  n = FALSE,
  keep_all = FALSE,
  con,
  base_url,
  my_schema
)

Arguments

cohort

tbl; reference to a table created in a user schema with a column called "person_id", and columns for start_date and end_date

concept_set_id

num; the number associated with the concept_set_id in ATLAS

start_date

the name of the start_date column in the cohort table (unquoted)

end_date

the name of the end_date column in the cohort table (unquoted)

name

chr; Name of column for indicator variable returned by function

min_n

dbl; the minimum number of occurrences per person to consider the indicator true

n

dbl; count the number of occurrences per person (will not include zeros)

keep_all

lgl; keep columns with information about the concept (e.g., concept name, id, etc.)

con

the connection object to ohdsilab

base_url

chr; the base url for ATLAS (not ohdsilab)

my_schema

chr; schema for the user

Value

a dataframe

Examples

if (FALSE) {
tobacco <- pull_concepts(con, cohort, concept_set_id = 1157, start_date = covariate_start_date,
 end_date = cohort_start_date, base_url = base_url, name = "tobacco", my_schema = my_schema
 )}